refactor(frontend): split useSettingsHandlers by domain (FE-6)
Extract 3 domain-scoped sub-hooks from the 1259-line god hook:
- useImportFolderHandlers (212 lines) — add/remove/scan folder management
- useBackupHandlers (167 lines) — create/restore/delete backup lifecycle
- useMetadataSourceHandlers (121 lines) — source toggle/test/reorder/credentials

useSettingsHandlers composes them with spread returns, keeping the
UseSettingsHandlersReturn interface identical — Settings.tsx is unchanged.
Net: 1259 → 936 lines in the main hook (−26%).
